autoremove env variable support to add user into db directly

+++ Invitarr.py
@@ -1,22 +1,25 @@
 #Copyright 2020 Sleepingpirate. 
-
+from os import environ
 import discord
 from discord.ext import commands
 import asyncio
 from plexapi.myplex import MyPlexAccount
 from discord import Webhook, AsyncWebhookAdapter
 import aiohttp
-import db as db
-
 
 # settings
 Discord_bot_token = '' 
-roleid =              # Role Id, right click the role and copy id.  
+roleid =            # Role Id, right click the role and copy id.  
 PLEXUSER = ''           # Plex Username
 PLEXPASS = ''           # plex password
 PLEX_SERVER_NAME = ''   # Name of plex server 
 Plex_LIBS = [] #name of the libraries you want the user to have access to.
 chan =  #Channel id of the channel you want to log emails and use -plexadd in. 
+auto_remove_user = environ.get('autoremoveuser') if environ.get('autoremoveuser') else False # auto remove user from plex and db if removed from the role
+
+if auto_remove_user:
+    print("auto remove user = True")
+    import db as db
 
 account = MyPlexAccount(PLEXUSER, PLEXPASS)
 plex = account.resource(PLEX_SERVER_NAME).connect()  # returns a PlexServer instance
@@ -70,7 +73,8 @@ class MyClient(discord.Client):
             print(email.content) #make it go to a log channel
             plexname = str(email.content)
             if plexadd(plexname):
-                db.save_user(after.display_name, email.content)
+                if auto_remove_user:
+                    db.save_user(after.display_name, email.content)
                 await asyncio.sleep(20)
                 await after.send('You have Been Added To Plex!')
                 secure = client.get_channel(chan)
@@ -78,17 +82,18 @@ class MyClient(discord.Client):
             else:
                 await after.send('There was an error adding this email address. Message Server Admin.')
         elif(role not in after.roles and role in before.roles):
-            try:
-                username = after.display_name
-                email = db.get_useremail(username)
-                plexremove(email)
-                deleted = db.delete_user(username)
-                if deleted:
-                    print("Removed {} from db".format(email))
-                else:
-                    print("Cannot remove this user from db.")
-            except:
-                print("Cannot remove this user from plex.")
+            if auto_remove_user:
+                try:
+                    username = after.display_name
+                    email = db.get_useremail(username)
+                    plexremove(email)
+                    deleted = db.delete_user(username)
+                    if deleted:
+                        print("Removed {} from db".format(email))
+                    else:
+                        print("Cannot remove this user from db.")
+                except:
+                    print("Cannot remove this user from plex.")
 
     async def on_message(self, message):
         secure = client.get_channel(chan)
@@ -108,6 +113,16 @@ class MyClient(discord.Client):
                     await message.channel.send('The email has been removed! {0.author.mention}'.format(message))
                 else:
                     message.channel.send('Error Check Logs! {0.author.mention}'.format(message))
+            if message.content.startswith('-adddb'):
+                mgs = message.content.replace('-adddb ','')
+                try:
+                    mgs = mgs.split(' ')
+                    email = mgs[0]
+                    username = mgs[1].replace('@', '').split('#')[0]
+                    db.save_user(username, email)
+                    await message.channel.send('The user {} has been added to db!'.format(mgs[1]))
+                except:
+                    print("Cannot add this user to db.")
 
 client = MyClient()
 client.run(Discord_bot_token)
